+++ Runtime/Scripts/Components/UIs/MiniMapUI.cs
@@ -1,21 +1,24 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using ParkMinPackages.Foundation.Constants;
 using ParkMinPackages.Foundation.Interfaces;
 using ParkMinPackages.Workflow.Default.Components.UIs;
 using ParkMinPackages.Workflow.Minimap.Components.Actors;
+using ParkMinPackages.Workflow.Minimap.Enums;
 using ParkMinPackages.Workflow.Minimap.Interfaces;
 using R3;
 using R3.Triggers;
 using Sirenix.OdinInspector;
 using UnityEngine;
+using UnityEngine.Serialization;
 using UnityEngine.UI;
 
 namespace ParkMinPackages.Workflow.Minimap
 {
 	[RequireComponent(typeof(RectMask2D))]
-	public class MiniMapUI : BasicUI, IMiniMapUI, IEnumerable<MiniMapMarkerUI>, IR3PostLateUpdatable
+	public class MiniMapUI : BasicUI, IMiniMapUI, IEnumerable<MiniMapElementUI>, IR3PostLateUpdatable
 	{
 		// - Public Methods -
 		public void Initialize(MiniMapCamera miniMapCamera) {
@@ -23,15 +26,19 @@ namespace ParkMinPackages.Workflow.Minimap
 				throw new ArgumentNullException(nameof(miniMapCamera));
 			if (MiniMapImage == null)
 				throw new NullReferenceException(nameof(MiniMapImage));
-			if (MiniMapImage.rectTransform.parent != transform)
-				throw new InvalidOperationException($"{nameof(MiniMapImage)} must be a direct child of {nameof(MiniMapUI)}.");
+			if (MapElementContainer == null)
+				throw new NullReferenceException(nameof(MapElementContainer));
+			if (OverlayElementContainer == null)
+				throw new NullReferenceException(nameof(OverlayElementContainer));
+
+			ValidateDirectChild(MiniMapImage.rectTransform, nameof(MiniMapImage));
+			ValidateDirectChild(MapElementContainer, nameof(MapElementContainer));
+			ValidateDirectChild(OverlayElementContainer, nameof(OverlayElementContainer));
 
 			MiniMapCamera.SpriteCaptureData captureData = miniMapCamera.CaptureSprite();
 			_captureData?.Dispose();
 			_captureData = captureData;
 			MiniMapImage.sprite = captureData.Sprite;
-			MiniMapImage.SetNativeSize();
-			MiniMapImage.rectTransform.SetAsFirstSibling();
 			SetView(new Vector3(captureData.WorldCenter.x, 0f, captureData.WorldCenter.y), Rotation, ViewWorldHeight);
 		}
 
@@ -57,7 +64,7 @@ namespace ParkMinPackages.Workflow.Minimap
 		}
 
 		public void RefreshView() {
-			if (_captureData == null || MiniMapImage == null || MiniMapImage.sprite == null)
+			if (_captureData == null || MiniMapImage == null || MiniMapImage.sprite == null || MapElementContainer == null || OverlayElementContainer == null)
 				return;
 
 			RectTransform frameRectTransform = (RectTransform)transform;
@@ -65,80 +72,85 @@ namespace ParkMinPackages.Workflow.Minimap
 			if (frameSize.x <= 0f || frameSize.y <= 0f)
 				return;
 
-			float uiSizePerWorldUnit = frameSize.y / ViewWorldHeight;
-			Vector2 imageSize = _captureData.WorldSize * uiSizePerWorldUnit;
-			Rect worldRect = _captureData.WorldRect;
+			_uiSizePerWorldUnit = frameSize.y / ViewWorldHeight;
+			_worldRect = _captureData.WorldRect;
+			_worldSize = _captureData.WorldSize;
+			Vector2 imageSize = _worldSize * _uiSizePerWorldUnit;
 			Vector2 normalizedCenter = new Vector2(
-				Mathf.InverseLerp(worldRect.xMin, worldRect.xMax, Center.x),
-				Mathf.InverseLerp(worldRect.yMin, worldRect.yMax, Center.z)
+				Mathf.InverseLerp(_worldRect.xMin, _worldRect.xMax, Center.x),
+				Mathf.InverseLerp(_worldRect.yMin, _worldRect.yMax, Center.z)
 			);
 
-			RectTransform imageRectTransform = MiniMapImage.rectTransform;
-			imageRectTransform.anchorMin = new Vector2(0.5f, 0.5f);
-			imageRectTransform.anchorMax = new Vector2(0.5f, 0.5f);
-			imageRectTransform.pivot = normalizedCenter;
-			imageRectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, imageSize.x);
-			imageRectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, imageSize.y);
-			imageRectTransform.anchoredPosition = Vector2.zero;
-			imageRectTransform.localRotation = Quaternion.Euler(0f, 0f, Rotation);
+			ApplyHierarchyView(imageSize, normalizedCenter);
+			_imageToOverlayMatrix = OverlayElementContainer.worldToLocalMatrix * MiniMapImage.rectTransform.localToWorldMatrix;
+			RefreshElements();
 		}
 
-		public Vector2 WorldToMiniMapPoint(Vector3 worldPosition) {
+		public Vector2 WorldToMapPoint(Vector3 worldPosition) {
 			if (IsInitialized == false)
 				throw new InvalidOperationException($"{nameof(MiniMapUI)} is not initialized.");
 
-			Rect worldRect = WorldRect;
-			Vector2 normalizedPoint = new Vector2(
-				(worldPosition.x - worldRect.xMin) / worldRect.width,
-				(worldPosition.z - worldRect.yMin) / worldRect.height
+			Vector2 normalizedPoint = WorldToNormalizedPoint(worldPosition);
+			Rect mapRect = MapElementContainer.rect;
+			return new Vector2(
+				Mathf.LerpUnclamped(mapRect.xMin, mapRect.xMax, normalizedPoint.x),
+				Mathf.LerpUnclamped(mapRect.yMin, mapRect.yMax, normalizedPoint.y)
 			);
-			RectTransform imageRectTransform = MiniMapImage.rectTransform;
-			Rect imageRect = imageRectTransform.rect;
-			Vector2 imagePoint = new Vector2(
+		}
+
+		public Vector2 WorldToOverlayPoint(Vector3 worldPosition) {
+			if (IsInitialized == false)
+				throw new InvalidOperationException($"{nameof(MiniMapUI)} is not initialized.");
+
+			Vector2 normalizedPoint = WorldToNormalizedPoint(worldPosition);
+			Rect imageRect = MiniMapImage.rectTransform.rect;
+			Vector3 imagePoint = new Vector3(
 				Mathf.LerpUnclamped(imageRect.xMin, imageRect.xMax, normalizedPoint.x),
-				Mathf.LerpUnclamped(imageRect.yMin, imageRect.yMax, normalizedPoint.y)
+				Mathf.LerpUnclamped(imageRect.yMin, imageRect.yMax, normalizedPoint.y),
+				0f
 			);
-			Vector3 worldPoint = imageRectTransform.TransformPoint(imagePoint);
-			return MarkerContainer.InverseTransformPoint(worldPoint);
+			Vector3 overlayPoint = _imageToOverlayMatrix.MultiplyPoint3x4(imagePoint);
+			return new Vector2(overlayPoint.x, overlayPoint.y);
 		}
 
-		public T CreateMarker<T>(Transform target, T markerPrefab) where T : MiniMapMarkerUI {
-			if (target == null)
-				throw new ArgumentNullException(nameof(target));
-			if (markerPrefab == null)
-				throw new ArgumentNullException(nameof(markerPrefab));
+		public T AddElement<T>(T elementPrefab, MiniMapElementLayer layer) where T : MiniMapElementUI {
+			if (elementPrefab == null)
+				throw new ArgumentNullException(nameof(elementPrefab));
 
-			T markerUI = Instantiate(markerPrefab, MarkerContainer);
-			markerUI.Initialize(target);
-			_markers.Add(markerUI);
-			markerUI.OnDestroyAsObservable().Subscribe(_ => _markers.Remove(markerUI)).AddTo(gameObject);
-			return markerUI;
+			RectTransform parent = GetElementParent(layer);
+			T elementUI = Instantiate(elementPrefab, parent);
+			RegisterElement(elementUI, layer);
+			return elementUI;
 		}
 
-		public void DestroyMarker(
-			MiniMapMarkerUI markerUI,
-			Action<MiniMapMarkerUI> destroyAction = null
+		public void DestroyElement(
+			MiniMapElementUI elementUI,
+			Action<MiniMapElementUI> destroyAction = null
 		) {
-			if (markerUI == null)
+			if (elementUI == null)
 				return;
 
-			_markers.Remove(markerUI);
+			RemoveElement(elementUI);
 			if (destroyAction != null)
-				destroyAction(markerUI);
+				destroyAction(elementUI);
 			else
-				Destroy(markerUI.gameObject);
+				Destroy(elementUI.gameObject);
 		}
 
-		public void ClearMarkers(
-			Action<MiniMapMarkerUI> destroyAction = null
+		public void ClearElements(
+			Action<MiniMapElementUI> destroyAction = null
 		) {
-			MiniMapMarkerUI[] markerUIs = _markers.ToArray();
-			for (int i = 0; i < markerUIs.Length; i++)
-				DestroyMarker(markerUIs[i], destroyAction);
+			MiniMapElementUI[] elementUIs = _elements.ToArray();
+			for (int i = 0; i < elementUIs.Length; i++)
+				DestroyElement(elementUIs[i], destroyAction);
 		}
 
-		public IEnumerator<MiniMapMarkerUI> GetEnumerator() {
-			return _markers.GetEnumerator();
+		public IEnumerable<T> GetElements<T>() where T : MiniMapElementUI {
+			return _elements.OfType<T>();
+		}
+
+		public IEnumerator<MiniMapElementUI> GetEnumerator() {
+			return _elements.GetEnumerator();
 		}
 
 		IEnumerator IEnumerable.GetEnumerator() {
@@ -154,17 +166,21 @@ namespace ParkMinPackages.Workflow.Minimap
 		{
 			get { return _miniMapImage; }
 		}
-		public RectTransform MarkerContainer
+		public RectTransform MapElementContainer
 		{
-			get { return _markerContainer; }
+			get { return _mapElementContainer; }
 		}
-		public IReadOnlyList<MiniMapMarkerUI> Markers
+		public RectTransform OverlayElementContainer
 		{
-			get { return _markers; }
+			get { return _overlayElementContainer; }
+		}
+		public IReadOnlyList<MiniMapElementUI> Elements
+		{
+			get { return _elements; }
 		}
 		public int Count
 		{
-			get { return _markers.Count; }
+			get { return _elements.Count; }
 		}
 		public Vector3 Center { get; private set; }
 		public float Rotation { get; private set; }
@@ -194,12 +210,7 @@ namespace ParkMinPackages.Workflow.Minimap
 		// - Handler -
 		protected override void Awake() {
 			base.Awake();
-			_markerContainer.anchorMin = Vector2.zero;
-			_markerContainer.anchorMax = Vector2.one;
-			_markerContainer.offsetMin = Vector2.zero;
-			_markerContainer.offsetMax = Vector2.zero;
-			_markerContainer.localScale = Vector3.one;
-			_markerContainer.SetAsLastSibling();
+			InitializeHierarchy();
 		}
 
 		protected override void Start() {
@@ -207,20 +218,8 @@ namespace ParkMinPackages.Workflow.Minimap
 			if (InitializeOnStart)
 				Initialize(_initializeOnStartMiniMapCamera);
 
-			List<MiniMapMarkerUI> markerUIs = new List<MiniMapMarkerUI>();
-			MarkerContainer.GetComponentsInChildren(true, markerUIs);
-			for (int i = 0; i < markerUIs.Count; i++) {
-				MiniMapMarkerUI markerUI = markerUIs[i];
-				if (markerUI.Target == null) {
-					Destroy(markerUI.gameObject);
-					continue;
-				}
-				if (_markers.Contains(markerUI))
-					continue;
-
-				_markers.Add(markerUI);
-				markerUI.OnDestroyAsObservable().Subscribe(_ => _markers.Remove(markerUI)).AddTo(gameObject);
-			}
+			RegisterHierarchyElements(MapElementContainer, MiniMapElementLayer.Map);
+			RegisterHierarchyElements(OverlayElementContainer, MiniMapElementLayer.Overlay);
 		}
 
 		protected void OnRectTransformDimensionsChange() {
@@ -231,62 +230,117 @@ namespace ParkMinPackages.Workflow.Minimap
 			if (IsInitialized == false)
 				return;
 
-			for (int i = _markers.Count - 1; 0 <= i; i--) {
-				MiniMapMarkerUI markerUI = _markers[i];
-				if (markerUI == null) {
-					_markers.RemoveAt(i);
-					continue;
-				}
-				if (markerUI.Target == null) {
-					markerUI.SetOutOfBounds(false);
-					markerUI.Hide();
-					continue;
-				}
-
-				Vector2 markerPoint = WorldToMiniMapPoint(markerUI.WorldPosition);
-				Rect containerRect = MarkerContainer.rect;
-				bool isOutOfBounds = containerRect.Contains(markerPoint) == false;
-				markerUI.SetOutOfBounds(isOutOfBounds);
-				if (markerUI.OutOfBounds == MiniMapMarkerUI.OutOfBoundsMode.Hide && isOutOfBounds) {
-					markerUI.Hide();
-					continue;
-				}
-
-				RectTransform markerRectTransform = markerUI.RectTransform;
-				Vector3 localPosition = markerRectTransform.localPosition;
-				markerRectTransform.localPosition = new Vector3(markerPoint.x, markerPoint.y, localPosition.z);
-				markerRectTransform.localRotation = markerUI.Rotation == MiniMapMarkerUI.RotationMode.WorldDirection
-					? Quaternion.Euler(0f, 0f, Rotation - markerUI.WorldYaw)
-					: Quaternion.identity;
-
-				if (markerUI.OutOfBounds == MiniMapMarkerUI.OutOfBoundsMode.Clamp) {
-					Bounds markerBounds = RectTransformUtility.CalculateRelativeRectTransformBounds(MarkerContainer, markerRectTransform);
-					Vector3 correction = Vector3.zero;
-					if (markerBounds.min.x < containerRect.xMin)
-						correction.x = containerRect.xMin - markerBounds.min.x;
-					else if (containerRect.xMax < markerBounds.max.x)
-						correction.x = containerRect.xMax - markerBounds.max.x;
-					if (markerBounds.min.y < containerRect.yMin)
-						correction.y = containerRect.yMin - markerBounds.min.y;
-					else if (containerRect.yMax < markerBounds.max.y)
-						correction.y = containerRect.yMax - markerBounds.max.y;
-					markerRectTransform.localPosition += correction;
-				}
-				markerUI.Show();
-			}
+			RefreshElements();
 		}
 
 		protected override void OnDestroy() {
-			_markers.Clear();
+			_elements.Clear();
 			_captureData?.Dispose();
 			_captureData = null;
 			base.OnDestroy();
 		}
 
-		// - Internals -
+		// - Private & Protected -
+		void RegisterHierarchyElements(RectTransform container, MiniMapElementLayer layer) {
+			if (container == null)
+				return;
+
+			List<MiniMapElementUI> elementUIs = new List<MiniMapElementUI>();
+			container.GetComponentsInChildren(true, elementUIs);
+			for (int i = 0; i < elementUIs.Count; i++)
+				RegisterElement(elementUIs[i], layer);
+		}
+		void RegisterElement(MiniMapElementUI elementUI, MiniMapElementLayer layer) {
+			if (elementUI == null || _elements.Contains(elementUI))
+				return;
+
+			elementUI.Init(this, layer);
+			_elements.Add(elementUI);
+			elementUI.OnDestroyAsObservable().Subscribe(_ => RemoveElement(elementUI)).AddTo(gameObject);
+		}
+		void RemoveElement(MiniMapElementUI elementUI) {
+			_elements.Remove(elementUI);
+		}
+		RectTransform GetElementParent(MiniMapElementLayer layer) {
+			return layer switch
+			{
+				MiniMapElementLayer.Map => MapElementContainer,
+				MiniMapElementLayer.Overlay => OverlayElementContainer,
+				_ => throw new ArgumentOutOfRangeException(nameof(layer), layer, null)
+			};
+		}
+		void InitializeHierarchy() {
+			InitializeRectTransform(MiniMapImage == null ? null : MiniMapImage.rectTransform);
+			InitializeRectTransform(MapElementContainer);
+			InitializeRectTransform(OverlayElementContainer);
+
+			if (MiniMapImage != null)
+				MiniMapImage.rectTransform.SetAsFirstSibling();
+			if (MapElementContainer != null)
+				MapElementContainer.SetSiblingIndex(1);
+			if (OverlayElementContainer != null)
+				OverlayElementContainer.SetAsLastSibling();
+		}
+		void InitializeRectTransform(RectTransform rectTransform) {
+			if (rectTransform == null)
+				return;
+
+			rectTransform.anchorMin = new Vector2(0.5f, 0.5f);
+			rectTransform.anchorMax = new Vector2(0.5f, 0.5f);
+			rectTransform.anchoredPosition = Vector2.zero;
+			rectTransform.localScale = Vector3.one;
+			rectTransform.localRotation = Quaternion.identity;
+		}
+		void ApplyHierarchyView(Vector2 imageSize, Vector2 normalizedCenter) {
+			RectTransform imageRectTransform = MiniMapImage.rectTransform;
+			imageRectTransform.pivot = normalizedCenter;
+			SetSize(imageRectTransform, imageSize);
+			imageRectTransform.anchoredPosition = Vector2.zero;
+			imageRectTransform.localScale = Vector3.one;
+			imageRectTransform.localRotation = Quaternion.Euler(0f, 0f, Rotation);
+
+			MapElementContainer.pivot = normalizedCenter;
+			SetSize(MapElementContainer, _worldSize);
+			MapElementContainer.anchoredPosition = Vector2.zero;
+			MapElementContainer.localScale = new Vector3(_uiSizePerWorldUnit, _uiSizePerWorldUnit, 1f);
+			MapElementContainer.localRotation = Quaternion.Euler(0f, 0f, Rotation);
+
+			OverlayElementContainer.pivot = normalizedCenter;
+			SetSize(OverlayElementContainer, imageSize);
+			OverlayElementContainer.anchoredPosition = Vector2.zero;
+			OverlayElementContainer.localScale = Vector3.one;
+			OverlayElementContainer.localRotation = Quaternion.identity;
+		}
+		void SetSize(RectTransform rectTransform, Vector2 size) {
+			rectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, size.x);
+			rectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, size.y);
+		}
+		void RefreshElements() {
+			for (int i = _elements.Count - 1; 0 <= i; i--) {
+				MiniMapElementUI elementUI = _elements[i];
+				if (elementUI == null) {
+					_elements.RemoveAt(i);
+					continue;
+				}
+
+				elementUI.RefreshView();
+			}
+		}
+		Vector2 WorldToNormalizedPoint(Vector3 worldPosition) {
+			return new Vector2(
+				(worldPosition.x - _worldRect.xMin) / _worldRect.width,
+				(worldPosition.z - _worldRect.yMin) / _worldRect.height
+			);
+		}
+		void ValidateDirectChild(RectTransform rectTransform, string name) {
+			if (rectTransform.parent != transform)
+				throw new InvalidOperationException($"{name} must be a direct child of {nameof(MiniMapUI)}.");
+		}
+
 		[Title(Headers.Required)]
 		[SerializeField, Required] Image _miniMapImage;
-		[SerializeField, Required] RectTransform _markerContainer;
+		[SerializeField, Required] RectTransform _mapElementContainer;
+		[FormerlySerializedAs("_markerContainer"), SerializeField, Required] RectTransform _overlayElementContainer;
 
 		[Title(Headers.Settings)]
 		[SerializeField] bool _initializeOnStart;
@@ -294,6 +348,10 @@ namespace ParkMinPackages.Workflow.Minimap
 		[SerializeField, Min(0.01f)] float _viewWorldHeight = 30f;
 
 		MiniMapCamera.SpriteCaptureData _captureData;
-		readonly List<MiniMapMarkerUI> _markers = new List<MiniMapMarkerUI>();
+		Rect _worldRect;
+		Vector2 _worldSize;
+		float _uiSizePerWorldUnit;
+		Matrix4x4 _imageToOverlayMatrix;
+		readonly List<MiniMapElementUI> _elements = new List<MiniMapElementUI>();
 	}
 }