+++ Runtime/Scripts/Enums/MiniMapElementLayer.cs
@@ -1,0 +1,8 @@
+namespace ParkMinPackages.Workflow.Minimap.Enums
+{
+	public enum MiniMapElementLayer
+	{
+		Map,
+		Overlay
+	}
+}